Correct the inclusion and usage of Eigen in CMakeLists.txt
Specifying `${EIGEN3_INCLUDE_DIR}` as an explicit catkin_package INCLUDE_DIR is not as good idea as putting it as a DEPENDS package. That will do the same thing and, moreover, is future-proof. Further, it also adds required compiler definitions, which you've forgot to do here (and I think lots of people would). This can have disastrous consequences if you change some flags that alter memory layout of Eigen objects.

Also, adding the include directory via a UseEigen3 file is more future-proof.